IDBlock::operator< instead of ltIDBlock.
Allow tasks to start in one thread and end later in another, as long as
it's on the same host and unambiguous.  That means table_tasks has two
thread_id fields: thread_start and thread_end.  Task class now also keeps
track of thread_id and path_id so it can stand out of context when checked
later.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1149 192cc142-4ac5-4de2-ad3d-7c493c50dd98