Added evtest, which interleaves pathids on a single thread.  It currently gets billed wrong, because reconciliation (dbfill) doesn't suspend accounting of path 1's task when we switch to path 2.
In Java bindings, more efficient pickling of ints.  Serializing an Integer
object results in a huge blob of data.

threadtest program now starts the task "catch-this" on one thread and ends
it on another, to test functionality that Macedon needs.

Each table builds an index on pathid column.  We lock the tables for
speed.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1153 192cc142-4ac5-4de2-ad3d-7c493c50dd98