All pathevents define a to_string function used by the DAG-click popup.
DAG no longer makes space for a caption, because we don't use it.

New window (also managed by glade) to display all of the events in a
thread that correspond to a single DAG node.

Don't display (in the DAG or pathtl) any malformed paths.  No more "-1"
click-to-crash.

Tweaked insert order for DAG, in hopes of making the display
chronological.  It's still wrong!


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1165 192cc142-4ac5-4de2-ad3d-7c493c50dd98