Moved get_threads into path.cc so both/all checkers can share it.
Changed GtkGraph to use "dot" (as an external program, not via library
calls) for its layout.

Changed GtkGraph to only perform layout when the graph is to be drawn, so
that hidden graphs don't pay that cost.

gtk_graph_add_edge returns the added edge.

GtkGraphEdge now has a series of control points, for splines.  Not that we
draw splines -- we just connect the dots.

Added the ability to graph metrics for all paths, in addition to all
instances of a task.  It's not fully implemented yet.  The title on the
Graph tab changes with whatever is graphed.

Normalize graph: suppress multiple edges between a pair of nodes; if
there's a directed edge in each direction, replace it with a single,
undirected edge.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1181 192cc142-4ac5-4de2-ad3d-7c493c50dd98