Support for invalidators in addition to validators and recognizers.
Removed "reverse" from the language -- it got cut from the implementation
back in March.

In pathcheck and pathview, changed the way paths are counted based on
matching validators and invalidators, and/or missing on their resource
limits.  This needs more work and probably should be shared code between
the two programs.

Bugfix in pathview: gtk_list_store_set takes char*s, not chars.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1191 192cc142-4ac5-4de2-ad3d-7c493c50dd98