New metric, busy-time = cputime/tdiff.  Used to express the notion that a path should (or shouldn't) spend its time computing as opposed to blocking.
Ranges can be floats, not just ints.

Task "real time" metric should be tdiff, not end-start.  For the moment,
they're the same, but maybe someday I'll fix the task-interruption code so
that tdiff <= end-start.

Bugfix: closing the taskpopup window hides it (like the dagpopup window)
rather than deleting it.  Fixes a segfault.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1193 192cc142-4ac5-4de2-ad3d-7c493c50dd98