Removed a few more references to SPLIT and JOIN, which have been gone for ages.
send() and recv() are now able to refer to a comma-separated list of
potential thread classes.  As in, a multicast interior node may send to
another interior node or to a leaf: send(Interior, Leaf);.

Updated pip.vim to understand send and recv.

In pathview, paths graph shows only paths visible in the paths list, so we
can filter by recognizer or by pattern.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1198 192cc142-4ac5-4de2-ad3d-7c493c50dd98