More graphable metrics for the paths graph: network bytes, threads, hosts, depth.  Still not latency.
Clicking on a point in the paths graph now displays (on stdout, ick!) some
of the pathstub info about that path.  To support this, the GtkPlot widget
now associates a user_data field with each point.

Renamed the top-level windows to start with "pathview_" so that .gtkrc
files can single them out.


git-svn-id: svn+ssh://aws/home/ubuntu/SVN/trunk/pip@1207 192cc142-4ac5-4de2-ad3d-7c493c50dd98